backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
alembic
pymysql
pydantic>=2.0.0
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
cryptography
jinja2
itsdangerous
